-PlyNG: -Staff Dashboard introduced. -Staff/Admin community models introduced. -Improved ACLs. -Introduced multiple dashboard types. -Community Staff management introduced into worldforge

+++ communities/community/forge_api_views.py
@@ -9,7 +9,7 @@ from ply import settings
 from ply.toolkit import vhosts,profiles,contexts,dynapages as dp_tools
 from dashboard.navigation import SideBarBuilder_dynamic
 from communities.community.models import CommunityAdmins,CommunityProfileDashboardRoles,CommunityStaff,CommunityDashboardType
-from communities.community.forms import CommunityStaffForm
+from communities.community.forms import CommunityStaffForm,CommunityAdminForm
 from communities.group.models import GroupMember
 from core.dynapages import models as dynapages
 from communities.profiles.models import ProfilePageNode
@@ -48,5 +48,36 @@ def delete_community_staff(request,staff):
     dbt = CommunityDashboardType.objects.get(type='staff')
     dashobj = CommunityProfileDashboardRoles.objects.get(profile=staffobj.profile, community=staffobj.community,type=dbt)
     dashobj.delete()
+    staffobj.delete()
+    return JsonResponse({"res": "ok"}, safe=False)
+
+
+
+@login_required
+def create_community_admin(request):
+    #  Ignore port:
+    vhost,community,context = contexts.default_context(request)
+    profile = profiles.get_active_profile(request)
+    is_admin = CommunityAdmins.objects.filter(community=community,profile=profile,active=True)
+    if (len(is_admin) < 1):
+        return render(request,"error-access-denied.html",{})
+    form = CommunityAdminForm(request.POST)
+    form.set_community(community)
+    if not form.is_valid():
+        return JsonResponse({"res": "err", "e": form.errors}, safe=False)
+    else:
+        form.save()
+        return JsonResponse({"res": "ok", "pk": form.instance.pk}, safe=False)
+
+
+
+@login_required
+def delete_community_admin(request,staff):
+    vhost,community,context = contexts.default_context(request)
+    profile = profiles.get_active_profile(request)
+    is_admin = CommunityAdmins.objects.filter(community=community,profile=profile,active=True)
+    if (len(is_admin) < 1):
+        return render(request,"error-access-denied.html",{})
+    staffobj = CommunityAdmins.objects.get(uuid=staff)
     staffobj.delete()
     return JsonResponse({"res": "ok"}, safe=False)
